pulled out utility function
