+++ shared/src/business/utilities/filterEmptyStrings.js
@@ -1,0 +1,22 @@
+/**
+ * recursively remove empty strings from an object
+ *
+ * @param {object} params object to be filtered
+ * @returns {object} filter object
+ */
+exports.filterEmptyStrings = params => {
+  const removeEmpty = obj => {
+    Object.keys(obj).forEach(key => {
+      if (obj[key] && typeof obj[key] === 'object') {
+        removeEmpty(obj[key]);
+      } else if (obj[key] === '') {
+        delete obj[key];
+      }
+    });
+  };
+
+  if (params) {
+    removeEmpty(params);
+  }
+  return params;
+};
